refactor(houseclean): split css plugin, parser and generator into different folders
